Add files via upload
The "heart.csv" dataset consists of patient information and medical parameters related to cardiovascular health. It is intended for use in heart disease prediction or analysis tasks. The dataset includes various features such as demographic information, lifestyle indicators, and clinical measurements, alongside a target variable indicating the presence or absence of heart disease.